Rework definition of pods.
This rephrases and expands the definition of pods in an attempt to be:
a) much more explicit about what technically constitutes a pod;
b) agnostic towards particular container implementations.

A change along these lines will be critical as Kubernetes starts to
support alternative container runtimes (for example, #2725).

In trying to be as clear as possible, this turned into a larger rewrite
than I had originally anticipated, but I am fairly confident that it is
still effectively a no-op in terms of the current Kubernetes
implementation.  But I stand correctable :-).

The change also introduces some other minor cleanup and drops the
link to the `lmctfy` paper, which seemed like a tenuous reference
(perhaps there is something more specific within the lmctfy project we
could link to instead?)