Assert pre and post conditions of all commands in test-cmd.sh
`hack/test-cmd.sh` contains a list of `kubectl` commands to test.
However, the effect of the execution of those commands was not systematically
checked. As a consequence, it was possible to break a command without breaking
those tests.

This commit surrounds every single `kubectl` command with assertions on
pre and post condition.