Only insert outputversion as a preferred mapping type if no error occurs
Some types will be looked up based on their internal type, which is passed
directly to RESTMapping(). RESTMapping() will search the entire version
list if no external types are passed, but OutputVersionMapper was bypassing
that because the default type was always present.

Change OutputVersionMapper to collapse empty versions, and when no external
type is specified, try with the preferred version and then try without.