Check for server certs instead of kubecfg certs in e2e tests
According to make-ca-cert.sh script, the server certs and kubecfg certs are
placed in the same directory on the master.  certs.go in e2e suite should work
fine if it checks for server certs instead of kubecfg's.

This is related to PR #5113, where kubecfg certs are not present on the master.