Don't use hostPort in integration test.
The integration test will fail if I check in my pending PR
to remove boundPods.  Kubernetes eventually does the right
thing, but getting the integration test to check expectations
is hard because the scheduling behavior is unpredictable.

The boundPods removal is needed to fix P1 bug and speeds up
the scheduler considerably.