service.go e2e: Cleanup ELBs from previous runs
Prior to attempting to create new ones, cleanup from previous runs.
Timeouts, 500s, etc. are possible here, and if they happen, you don't
want to die forever.

Along the way: Remove the timeout, it was clearly copied from the
previous function and is actually an anti-pattern that needs to be
fixed after discovering it doesn't play well with defers.

Fixes #5722