Fix confusing use of "comprise"
The word "comprise" means "be composed of" or "contain" so
"applications comprised of multiple containers" would mean
"applications composed of of multiple containers" or "applications
contained of multiple containers" which is confusing.

I understand that this is nitpicking and that "comprise" has a new
meaning which is the opposite of its original definition just like
how "literally" now means "figuratively" to some people. However, I
believe that clarity is of utmost importance in technical
documentation which is why I'm proposing this change.