Allow marking node unschedulable in case of kubelet pushing node updates to master
kube-controller-manager has '--sync_node_status' flag and when this is set,
node controller probes kubelet for updates and when this is unset(default),
kubelet pushes updates to master. Ability to mark node unschedulable works
in the former case but is broken in the latter case. This change will fix the issue.