Kubelet: update mirror pods on static pods changes
This change checks whether the observed mirror pod is semantically equal to the
static pod. If not, it sends a request to the API server to update the mirror
pod. The update may be performed multiple times until it is reflected through
kubelet's pod watch. This is expected and should cause no harm.

Currently, because kubelet does not validate updates for static pods, it could
happen that the update request would be rejected by the API server. In this
case, we surface the error through events.