Support the absence of a source RC
When performing a rolling update, scale up the new RC immediately
in the absence of a source RC. This is to handle initial deployments
and any other case where the source RC either never existed or was
otherwise removed from the system. With no source RC, there's no reason
to do a gradual transition.