kubelet: support backoff on container restarts
When a container is stuck in a crash loop, kubelet would waste resources trying
to restart it repeatedly. We want to reduce the impact of these unhealthy
containers.

This change instructs per-pod worker to back off on unexpected restarts, which
excludes user-made container changes.
  * Upon an unexpected restart, a pod worker would double the current backoff
    interval, until it reaches the maximum interval.
  * If no unexpected restart has occured in the past 5 minutes, the interval
    would be reset. This ensures that we do not carry over the effects of
    ancient restarts.

For containers with failure frequency lower than once per 5 minutes, the impact
of source usuage is minimal. There is no need to back off.