Add short sleep after server startup.
The container is 'running', when its main process is started. That does not
mean that the server itself is running. Kubernetes may try to mount the fs
even before the server container has exported it. This may lead to long
timeouts and slow test. (Kubernetes _will_ retry to mount the FS if the first
mount has failed, so it's just about the timeouts).