Add `kubectl get -f file.json --latest=false`
Allows someone to convert an object on disk to the newest format
without that resource being on the server.

For instance, read all of the JSON files in the directory "foo"
and convert them to version v1beta3

    $ kubectl get -f foo/ --latest=false -o json > newfile.json

Makes it much easier to convert old resources to new resources.

Adds -f to kubectl get as well.