Workaround for race between monit and init
We start e.g. kube-proxy from /etc/init.d/,
and we use monit to restart it if it fails.

On a reboot though, init starts monit, and monit notices that
kube-proxy is not (yet) running, and starts kube-proxy.  But
init will also start kube-proxy.

When we start two copies of kube-proxy, the pid file is no longer correct,
and monit then goes into a restart loop.

Workaround: have monit wait on boot before checking anything.

This is only a workaround, not a fix, because there is still a race if the
system is running really slowly.  The real fix is probably to have monit
be responsible for starting the process, not init.

Bug #8742