Use --service-node-port-range flag value
Prevents "Node port range unspecified. Defaulting to 30000-32766." log message.
I believe the ServiceNodePortRange.Size field is always zero as it is set nowhere. As a resulte the --service-node-port-range flag value is ignored. So it is reasonable to compare ServiceNodePortRange with nil as it is done for ServiceClusterIPRange a few lines before.