kubelet: Lookup Cluster DNS Service if not specified on command line
This avoids the need to meticulously:

 a) Allocate and choose a cluster DNS IP
 b) Reconfigure kubelet --cluster-dxxxx startup args on all nodes

We lookup the cluster DNS IP and cluster Domain from the Services in
Kubernetes. This happens if a kubelet did not receive --cluster-dns
and --cluster-domain arguments on the command line.

The DNS service must be called 'kubernetes-dns' and be in the 'default'
namespace to be found. The service's ClusterIP is used as the DNS IP
address. The following annotation on the service is used as the domain name:

    kubernetes.io/cluster-domain = example.local

This eases deployment of new kubernetes nodes, by removing variables
that needs to be configured when deploying nodes.

This is a backwards compatible change.