Take systemd service management out of Salt
We use our own script instead, and call out to it from salt.

The version of Salt that we're running seems to have a very
long list of bugs with systemd, particularly on Ubuntu 15.04.
It does not look like it would actually fix things if we upgraded
to the latest Salt, and experience tells us that will end up breaking
more stuff.

So, use salt to cmd.run the actual things we want, and create a simple
script so we keep this logic in one place.

Fixes #9892