Validate endpoints in e2e tests by pod UID and port
Before this patch endpoints were validated by container IP and port.
Depending on the endpoint controller logic neither of the two must match for a
valid endpoint (e.g. in a Mesos setup).

This patch checks that the endpoint targetRef points to the right pod by UID,
instead of comparing IPs.

A later patch will make sure the
compared port is the actual container port, not the host port.

/xref mesosphere/kubernetes-mesos#365