Announce webserver port in NodePort e2e test
In k8s's default networking model container ports in pod specs are not necessary
in order to get an endpoint for a given service via the endpoint controller. In
other setups like Mesos this information is essential to allow the creation of
necessary port forwardings.

Fixes mesosphere/kubernetes-mesos#365: "Services should be able to create a functioning NodePort service",