updated kubernetes daemon docs
These docs are not autogenerated. They are generated from specially formated
markdown files that need to be updated when flags are changed but nobody knows
about this so it is not enforced. Thus the man page generation is not useful.
For the sake of having up to date docs up to date for v1, I am manually udating
the markdown files and removing the md2man generation. After v1 we need to gen
this man pages with the rest of our generated docs in a pre commit hook so that
they are not able to become out of date. We can do this by converting our daemons
to use cobra cmd like kubectl.