Updated command to launch etcd docker container in mesos documentation
The command to run the etcd docker didn't work for me. I kept getting a `connection refused` error when trying to CURL all the exposed ports of the container.

I was able to get it to work properly by referencing the command to launch the etcd container by using the command in the etcd documentation: https://github.com/coreos/etcd/blob/master/Documentation/docker_guide.md

This PR updates the docs so that the etcd container launches correctly.