Vagrant: Allow env override of IP addresses
MASTER_IP and MINION_IP_BASE are hard-coded in vagrant's
config-default.sh, and the values correspond to virtualbox's default
subnet.  On hosts that have both virtualbox and another provider
installed, attempting to deploy kubernetes with the non-virtualbox
provider is likely to result in broken networking.  This change allows
the addresses to be overridden via the environment so that more
appropriate values can be used.