Error: kubectl --all-namespaces get {node|ev|cs}
Report an error if someone asks for --all-namespaces
when getting a thing that is not namespaced.

This is in preparation for a subsequent commit which prints namespace
as its own column.

Restructured test to expect an error for non-namespaced things.
Dropped the part where it was trying to test that not printing
namespace didn't contain namespace.  Some other test can cover that.