Add cherry_pick_pull.sh <pr-number> <branch>
This script checks out a branch based on <branch>, cherry picks pull
request <pr> into it commit by commit, then instructs the user on how
to propose that branch as a PR to the branch.

For bonus points, this could be integrated with something like the
"hub" tool (https://github.com/github/hub) and be made about 2x more
automated.

Fixes #10730