Remove "Notes" column.
This mostly stated what version the code was last tested at.
This was useful when kubernetes was changing rapidly.
However, this was not updated reliably.  Also, simply using
a particular binary version does not assure users that
the distribution provides a conformant implementation.
The conformance test does provide that, and I'm hoping
that more distros will use that in the future.
Therefore, I've deleted the Notes column in favor of the
Conforms column.